refactor(enrichment): move RecordEnrichedPayload to schemas.py (Pair C P3)
- Colocate promotion boundary model with EnrichedJobProfile per docs/guides/Emilio.md
- Remove top-level enrichment/types.py to avoid confusion with enrichment.dedup.types
- Validators raise ValueError so pydantic surfaces ValidationError in tests
